+++ HISTORY.rst
@@ -3,7 +3,7 @@ History
 =======
 
 0.1.7 (2020-01-20)
-----------------
+------------------
 
 * Add support for Python 3.8 and Django 2.2
 * Fix generating of translations